+++ IL2-SR-Client/Settings/GlobalSettingsStore.cs
@@ -74,6 +74,7 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Client.Settings
 
         DisableWindowVisibilityCheck ,
         PlayConnectionSounds,
+        NeutralLobbyMusicEnabled,
 
         RequireAdmin,
 
@@ -395,6 +396,7 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Client.Settings
 
             {GlobalSettingsKeys.DisableWindowVisibilityCheck.ToString(), "true"},
             {GlobalSettingsKeys.PlayConnectionSounds.ToString(), "true"},
+            {GlobalSettingsKeys.NeutralLobbyMusicEnabled.ToString(), "true"},
 
             {GlobalSettingsKeys.RequireAdmin.ToString(),"true" },